+++ CHANGELOG.md
@@ -11,6 +11,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Fixed
 ### Security
 
+## [1.1.5] - 08/03/19
+### Added
+- Tests for configs
+### Changed
+- Error messages
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.1.4] - 08/03/19
 ### Added
 ### Changed